Promote the reference-derived G1 to robots/g1.py
The reshaped fork of g1_molmo's own components/robot.py becomes the codebase's
G1Robot; the earlier independent rewrite moves aside to
robots/g1_old_reference.py. Both moved with git mv so history follows.

Imports repointed by which implementation each consumer actually wants -- the
two are NOT interchangeable (constructors differ: (mj_data, exp_config) vs
(model, data, ...)), so this is not a blind rename:
  - configs/robot_configs.py, component_tests/test_g1_robot.py
    -> g1_old_reference (what G1Config/the native pipeline construct)
  - g1_molmo_port/env_g1ms.py, components/robot_g1ms.py shim, nav_demo.py
    -> robots.g1 (the reference-derived one)
No `g1_ported` references remain.

Both module docstrings now state which is which, how they differ, and records
the three known behavioural gaps in the old one (arm actfrcrange/dof_damping,
action noise on the IK command, non-persisted height smoothing).

Also clears two pre-existing ruff SIM105 findings in env_g1ms.py that
pre-commit surfaces now that the file is touched.

Verified: pick rollout byte-identical to baseline (58/58; steps=2338
sim_time=12.04s success=True) and native test_g1_robot.py still passes, so
neither pipeline regressed.
