+++ molmo_spaces/robots/g1.py
@@ -1,147 +1,364 @@
-from typing import TYPE_CHECKING, Any
+"""The G1 humanoid robot.
+
+Descends from g1_molmo's own components/robot.py -- started as a byte-identical
+fork of it, then reshaped onto molmo_spaces/robots/abstract.py's Robot
+interface (namespace, robot_view, kinematics, parallel_kinematics, controllers)
+one step at a time, each step verified byte-identical against the gold pick
+rollout. That provenance is why this, rather than the earlier independent
+rewrite, is the implementation the codebase uses: its grasp behaviour is
+verifiably the reference's, not a reimplementation of it.
+
+The superseded independent rewrite is kept alongside as
+`molmo_spaces/robots/g1_old_reference.py` -- it is what `G1Config` and the
+native pick/nav pipeline still construct, and the two are NOT drop-in
+interchangeable (different constructors: `(mj_data, exp_config)` there,
+`(model, data, ...)` here). See that module's docstring for the differences.
+
+Verification harnesses:
+  scripts/g1_molmo_port_comparison/generate_ported_rollout.py  (grasp, vs gold)
+  scripts/g1_molmo_port_comparison/nav_demo.py                 (navigation)
+"""
+
+import contextlib
+import re
+from typing import TYPE_CHECKING
 
 import mink
 import mujoco
 import numpy as np
-from mujoco import MjData, MjSpec, mjtEq, mjtObj
-from scipy.spatial.transform import Rotation as R
+from scipy.spatial.transform import Rotation as _R
 
-from molmo_spaces.controllers.abstract import Controller
-from molmo_spaces.controllers.g1_walk import (
-    _DEFAULT_HEIGHT_CMD,
-    NUM_TARGET_DIMS,
-    G1WalkController,
-)
-from molmo_spaces.controllers.joint_pos import JointPosController
 from molmo_spaces.kinematics.mujoco_kinematics import MlSpacesKinematics
 from molmo_spaces.kinematics.parallel.dummy_parallel_kinematics import DummyParallelKinematics
+from molmo_spaces.molmo_spaces_constants import ASSETS_DIR
 from molmo_spaces.robots.abstract import Robot
-from molmo_spaces.robots.robot_views.g1_view import (
-    ARM_JOINT_SUFFIXES,
-    HOLO_BASE_TARGET_BODY_NAME,
-    LEGS_WAIST_JOINT_SUFFIXES,
-    G1RobotView,
-)
 from molmo_spaces.utils.linalg_utils import normalize_ang_error
-
-# Velocity clamps for the AStarPlannerPolicy waypoint -> G1WalkController command
-# bridge (see G1Robot._waypoint_to_velocity_target). Not from the reference (which
-# doesn't have this bridge) -- chosen to match the forward speed already validated
-# empirically in the standing/walking smoke test.
-_MAX_LINEAR_VEL = 0.5
-_MAX_YAW_RATE = 0.5
-# G1WalkController switches from its walk policy to its (non-turning, non-walking)
-# stand policy whenever norm(cmd) < 0.05 (see g1_walk.py). A pure proportional law
-# decays toward zero as the error shrinks, so once any residual error's raw command
-# drops below that switch threshold, it gets stuck standing just short of the goal --
-# confirmed empirically via rotate() plateauing ~6-7deg short of target. Once the raw
-# error exceeds _VELOCITY_DEADBAND, floor the command at _MIN_* (comfortably above
-# 0.05) instead of letting it decay through the stand/walk switch point.
-_MIN_LINEAR_VEL = 0.15
-_MIN_YAW_RATE = 0.15
-_VELOCITY_DEADBAND = 0.08
-# Turn-then-drive gate (see _waypoint_to_velocity_target): only suppress translation
-# while heading error is *large*. Gating on "any nonzero yaw_rate" instead (i.e. the
-# full _VELOCITY_DEADBAND) was too strict -- G1WalkController's yaw tracking has its
-# own residual convergence ceiling around 15deg (confirmed via rotate()), so a target
-# heading landing inside that dead zone could never fully clear the gate, permanently
-# blocking translation even though the heading is already close enough to walk toward
-# the waypoint. Confirmed empirically: nav_to() stalling flat (no distance progress
-# for 30+ steps) at exactly this kind of near-but-not-exact heading alignment.
-_YAW_GATE_THRESHOLD = np.radians(30)
-
-# kinematics_wbc() constants -- exact port of g1_molmo's own GraspPolicy/
-# G1Controller mink whole-body IK setup (see molmo_spaces.g1_molmo_port.
-# components.robot_g1ms.G1Robot.kinematics_wbc, which prototypes this same
-# split onto a standalone G1Robot before this move). Torso-tilt safety
-# envelope kept in sync with g1_server deploy clips.
-_WBC_HEIGHT_MIN, _WBC_HEIGHT_MAX = 0.35, 0.793
-_WBC_WAIST_JOINTS = ("waist_yaw_joint", "waist_roll_joint", "waist_pitch_joint")
-_WBC_TORSO_TILT_RANGES = [
-    ("waist_yaw_joint", -0.5, 0.5),
-    ("waist_roll_joint", -0.4, 0.4),
-    ("waist_pitch_joint", -0.1, 0.4),
-]
-
-
-def _wbc_hand_arm_joints(side: str) -> tuple[str, ...]:
-    return tuple(f"{side}_{suffix}" for suffix in ARM_JOINT_SUFFIXES)
-
-
-_WBC_HAND_ARM_JOINTS = {side: _wbc_hand_arm_joints(side) for side in ("left", "right")}
-_WBC_HAND_SITE = {side: f"{side}_grasp" for side in ("left", "right")}
 
 if TYPE_CHECKING:
     from molmo_spaces.configs.abstract_exp_config import MlSpacesExpConfig
-    from molmo_spaces.configs.robot_configs import BaseRobotConfig
+
+XML_PATH = str(ASSETS_DIR / "robots/g1/g1_dex.xml")
+PREFIX = "robot_0/"
+ROOT_BODY = f"{PREFIX}pelvis"
+STANDING_HEIGHT = 0.75
+
+# Waypoint -> base-velocity bridge (see G1Robot.waypoint_to_velocity_target).
+# Not from the reference stack, which emits base velocities directly; these are
+# native molmo_spaces/robots/g1.py's own empirically-validated values, kept in
+# sync so a navigation policy drives either implementation identically.
+MAX_LINEAR_VEL = 0.5
+MAX_YAW_RATE = 0.5
+MIN_LINEAR_VEL = 0.15
+MIN_YAW_RATE = 0.15
+VELOCITY_DEADBAND = 0.08
+YAW_GATE_THRESHOLD = np.radians(30)
+
+# Body-frame +x offset (m) shifting reported "robot xy" forward of the pelvis to better
+# match the footprint center. Applied by get_xy()/set_pose() and the controller's _xy().
+PELVIS_FORWARD_OFFSET = 0.05
+
+# Dex gripper: positive qpos closes the fingers, negative opens.
+GRIPPER_OPEN = -0.0222
+GRIPPER_CLOSED = 0.0245
+
+# 30 joints = 12 legs + 3 waist + 14 arms + 1 right gripper (Joint2_1 is <equality>-coupled).
+JOINT_NAMES = [
+    "left_hip_pitch_joint",
+    "left_hip_roll_joint",
+    "left_hip_yaw_joint",
+    "left_knee_joint",
+    "left_ankle_pitch_joint",
+    "left_ankle_roll_joint",
+    "right_hip_pitch_joint",
+    "right_hip_roll_joint",
+    "right_hip_yaw_joint",
+    "right_knee_joint",
+    "right_ankle_pitch_joint",
+    "right_ankle_roll_joint",
+    "waist_yaw_joint",
+    "waist_roll_joint",
+    "waist_pitch_joint",
+    "left_shoulder_pitch_joint",
+    "left_shoulder_roll_joint",
+    "left_shoulder_yaw_joint",
+    "left_elbow_joint",
+    "left_wrist_roll_joint",
+    "left_wrist_pitch_joint",
+    "left_wrist_yaw_joint",
+    "right_shoulder_pitch_joint",
+    "right_shoulder_roll_joint",
+    "right_shoulder_yaw_joint",
+    "right_elbow_joint",
+    "right_wrist_roll_joint",
+    "right_wrist_pitch_joint",
+    "right_wrist_yaw_joint",
+    "right_Joint1_1",
+]
+
+ACTUATOR_NAME_MAP = {
+    "right_Joint1_1": "right_grip",
+}
+
+# Left arm values are near the gravity-settled hanging pose (the arm is unactuated).
+_DEFAULT_QPOS_PATTERNS = {
+    r"left_shoulder_pitch_joint": 0.212,
+    r"left_shoulder_roll_joint": -0.017,
+    r"left_shoulder_yaw_joint": 0.062,
+    r"left_elbow_joint": 1.216,
+    r"left_wrist_roll_joint": 0.005,
+    r"left_wrist_pitch_joint": 0.258,
+    r"left_wrist_yaw_joint": 0.006,
+    r".*_hip_pitch_joint": -0.312,
+    r".*_knee_joint": 0.669,
+    r".*_ankle_pitch_joint": -0.363,
+    r"right_elbow_joint": -0.2,
+    r"right_shoulder_roll_joint": -0.2,
+    r"right_shoulder_pitch_joint": 0.2,
+    r"right_Joint1_1": GRIPPER_OPEN,
+    r"right_Joint2_1": GRIPPER_OPEN,
+}
+
+
+def _resolve_defaults():
+    result = np.zeros(len(JOINT_NAMES), dtype=np.float64)
+    for i, name in enumerate(JOINT_NAMES):
+        for pat, val in _DEFAULT_QPOS_PATTERNS.items():
+            if re.fullmatch(pat, name):
+                result[i] = val
+                break
+    return result
+
+
+DEFAULT_QPOS = _resolve_defaults()
+
+
+def _is_floor_geom(name: str) -> bool:
+    name = (name or "").lower()
+    return (
+        name == "floor" or name.startswith("room|") or name.startswith("room_") or "floor" in name
+    )
+
+
+class G1RobotView:
+    """Pose/contact/visibility helpers for the G1 robot, split out of G1Robot
+    to match molmo_spaces/robots/robot_views/g1_view.py's G1RobotView --
+    the interface molmo_spaces/robots/abstract.py's Robot.robot_view exposes.
+    Same underlying code as before, just accessed through
+    `robot.robot_view.get_xy()` etc. instead of directly on G1Robot (which
+    keeps its own same-named methods as thin pass-throughs to this class, so
+    existing call sites are unaffected by this split).
+    """
+
+    def __init__(
+        self, model, data, body_id, freejoint_id, cam_ids, dof_ids, namespace: str = PREFIX
+    ):
+        self.model = model
+        self.data = data
+        self._body_id = body_id
+        self._freejoint_id = freejoint_id
+        self._cam_ids = cam_ids
+        self._dof_ids = dof_ids
+        self._namespace = namespace
+        self._renderer = None
+
+    def set_pose(self, xy, yaw, z=STANDING_HEIGHT):
+        qadr = self.model.jnt_qposadr[self._freejoint_id]
+        c, s = np.cos(yaw), np.sin(yaw)
+        px = xy[0] - c * PELVIS_FORWARD_OFFSET
+        py = xy[1] - s * PELVIS_FORWARD_OFFSET
+        self.data.qpos[qadr : qadr + 3] = [px, py, z]
+        self.data.qpos[qadr + 3 : qadr + 7] = _R.from_euler("z", yaw).as_quat(scalar_first=True)
+
+    def zero_velocities(self):
+        dadr = self.model.jnt_dofadr[self._freejoint_id]
+        self.data.qvel[dadr : dadr + 6] = 0.0
+        for i in range(len(JOINT_NAMES)):
+            self.data.qvel[self._dof_ids[i]] = 0.0
+
+    def get_xy(self):
+        return self.data.xpos[self._body_id, :2].copy()
+
+    def get_yaw(self):
+        quat = self.data.xquat[self._body_id]
+        return _R.from_quat(quat[[1, 2, 3, 0]]).as_euler("xyz")[2]
+
+    def pelvis_height(self):
+        qadr = self.model.jnt_qposadr[self._freejoint_id]
+        return float(self.data.qpos[qadr + 2])
+
+    def has_bad_contacts(self):
+        m, d = self.model, self.data
+        for i in range(d.ncon):
+            con = d.contact[i]
+            g1, g2 = int(con.geom1), int(con.geom2)
+            n1 = mujoco.mj_id2name(m, mujoco.mjtObj.mjOBJ_BODY, m.geom_bodyid[g1]) or ""
+            n2 = mujoco.mj_id2name(m, mujoco.mjtObj.mjOBJ_BODY, m.geom_bodyid[g2]) or ""
+            r1, r2 = n1.startswith(self._namespace), n2.startswith(self._namespace)
+            if r1 == r2:
+                continue
+            scene_geom = g2 if r1 else g1
+            if _is_floor_geom(mujoco.mj_id2name(m, mujoco.mjtObj.mjOBJ_GEOM, scene_geom) or ""):
+                continue
+            return True
+        return False
+
+    def check_object_visibility(self, body_id, threshold=0.00002):
+        # threshold = minimum fraction of the 224x224 segmentation frame the
+        # object must occupy in at least one camera. ~0.00002 (~1px) restores
+        # the original presence check: a hard 0.2% (~100px) gate is unsatisfiable
+        # for small/distant objects at far nav spawns, so placement fails after
+        # its retry budget and crashes rollout workers. Keep this near 1px.
+        if not self._cam_ids:
+            return True
+        if self._renderer is None:
+            # procthor scenes can exceed the default max_geom.
+            self._renderer = mujoco.Renderer(
+                self.model, 224, 224, max_geom=max(20000, self.model.ngeom * 4)
+            )
+        m = self.model
+        geom_ids = set()
+        stack = [body_id]
+        while stack:
+            bid = stack.pop()
+            for gid in range(m.ngeom):
+                if m.geom_bodyid[gid] == bid:
+                    geom_ids.add(gid)
+            for cbid in range(m.nbody):
+                if m.body_parentid[cbid] == bid and cbid != bid:
+                    stack.append(cbid)
+
+        for cam_id in self._cam_ids:
+            try:
+                self._renderer.update_scene(self.data, cam_id)
+                self._renderer.enable_segmentation_rendering()
+                seg = self._renderer.render()
+                self._renderer.disable_segmentation_rendering()
+            except IndexError:
+                return True  # segid overflow — treat as visible to avoid crashing reset.
+            seg0 = seg[:, :, 0]
+            vis = np.isin(seg0, list(geom_ids)).sum()  # object's visible pixels
+            if vis / seg0.size >= threshold:  # fraction of the frame
+                return True
+        return False
+
+    def close(self):
+        """Free the lazily-created visibility renderer (and its GL/EGL context).
+        Must be called before dropping the robot on scene reload — otherwise the
+        renderer's framebuffer leaks on the render GPU because __del__-based EGL
+        teardown is unreliable, so VRAM creeps to OOM across reloads."""
+        r = getattr(self, "_renderer", None)
+        if r is not None:
+            with contextlib.suppress(Exception):
+                r.close()
+            self._renderer = None
+
+
+# IK constants/joint groups mirroring agents/policy_g1ms.py's module-level
+# HEIGHT_MIN/HEIGHT_MAX, _WAIST, _hand()/_HANDS -- duplicated here (not
+# imported) because policy_g1ms.py imports FROM this module's predecessor
+# (components/robot.py) and would create a cycle; kept byte-identical to
+# those definitions since G1Robot.kinematics below is a verbatim relocation
+# of GraspPolicy._solve_ik, which depends on matching them exactly.
+HEIGHT_MIN, HEIGHT_MAX = 0.35, 0.793
+IK_DT = 1e-2
+IK_HEIGHT_DAMPING = 5e5
+WAIST_JOINTS = ("waist_yaw_joint", "waist_roll_joint", "waist_pitch_joint")
+
+
+def _hand_joints(side):
+    return tuple(
+        f"{side}_{j}"
+        for j in (
+            "shoulder_pitch_joint",
+            "shoulder_roll_joint",
+            "shoulder_yaw_joint",
+            "elbow_joint",
+            "wrist_roll_joint",
+            "wrist_pitch_joint",
+            "wrist_yaw_joint",
+        )
+    )
+
+
+HAND_ARM_JOINTS = {side: _hand_joints(side) for side in ("left", "right")}
+HAND_SITE = {side: f"{side}_grasp" for side in ("left", "right")}
 
 
 class G1Robot(Robot):
-    """G1 humanoid robot (Phase 3: whole-body walking via `G1WalkController`).
+    def __init__(
+        self,
+        model,
+        data,
+        env=None,
+        low_level=None,
+        namespace: str = PREFIX,
+        xml_path: str = XML_PATH,
+        exp_config: "MlSpacesExpConfig | None" = None,
+    ):
+        # Robot.__init__ wants (mj_data, exp_config) and only stores them; this
+        # class is constructed with an explicit (model, data) pair by the
+        # reference stack, so pass `data` through and keep self.model/self.data
+        # as the names the ported code already reads.
+        super().__init__(data, exp_config)
+        self.model = model
+        self.data = data
+        self._env = env
+        # Robot ABC solver properties, built lazily (see `kinematics`).
+        self._kinematics = None
+        self._parallel_kinematics = None
+        # Namespace/asset path are per-instance rather than the module-level
+        # PREFIX/XML_PATH the reference stack hardcoded, matching how every
+        # other molmo_spaces Robot takes these from its BaseRobotConfig
+        # (robot_namespace / get_robot_xml_path). The module constants remain
+        # the defaults, so the reference stack's own call sites are unchanged.
+        self._namespace = namespace
+        self._xml_path = xml_path
+        self._root_body = f"{namespace}pelvis"
 
-    The combined `legs_waist` move group (see `G1LegsWaistGroup`) is driven by
-    `G1WalkController`, a ported whole-body walking controller (WBC): a Python
-    PD-torque law plus an ONNX policy selecting between standing and walking
-    gaits, conditioned on a commanded [vx, vy, yaw_rate, height, waist] target
-    (see `G1WalkController.set_target`). This requires the `legs_waist`
-    actuators to behave as raw torque passthroughs rather than the MJCF's own
-    position-PD gains (see `apply_control_overrides`).
-
-    Both arms and the right gripper stay on plain JointPosControllers -- the
-    MJCF's own tuned PD actuator gains (biastype="affine") do the rest. The
-    base has no actuators; the pelvis is a free-floating body whose dynamics
-    MuJoCo integrates directly.
-    """
-
-    def __init__(self, mj_data: MjData, exp_config: "MlSpacesExpConfig") -> None:
-        super().__init__(mj_data, exp_config)
-
-        # Matches g1_molmo's own components/controller.py set_env() (`m.opt.
-        # timestep = 0.005`) -- see G1Config.physics_timestep's docstring for
-        # why. Applied here (before any physics stepping happens for this
-        # scene/robot) rather than left at the scene's own default.
-        physics_timestep = self.exp_config.robot_config.physics_timestep
-        if physics_timestep is not None:
-            mj_data.model.opt.timestep = physics_timestep
-
-        self._namespace = self.exp_config.robot_config.robot_namespace
-        self._use_holo_base = self.exp_config.robot_config.use_holo_base
-        self._robot_view = G1RobotView(mj_data, self.namespace, use_holo_base=self._use_holo_base)
-        self._kinematics = MlSpacesKinematics(self.exp_config.robot_config)
-        self._parallel_kinematics = DummyParallelKinematics(
-            self.exp_config.robot_config, self._kinematics
-        )
-        # Holo-base mode only: pending mocap-target ctrl for the "base" move group,
-        # applied in compute_control(). None means "no explicit waypoint command
-        # yet" -- falls back to base.noop_ctrl (the pelvis's live current pose,
-        # recomputed fresh every tick), so a freshly (re)placed robot never gets
-        # yanked toward a stale target left over from an earlier episode/pose.
-        self._pending_base_ctrl: np.ndarray | None = None
-
-        if self._use_holo_base:
-            legs_waist_controller = JointPosController(self.robot_view.get_move_group("legs_waist"))
-        else:
-            legs_waist_controller = G1WalkController(
-                self.robot_view.get_move_group("legs_waist"),
-                base_move_group=self.robot_view.get_move_group("base"),
-                left_arm_move_group=self.robot_view.get_move_group("left_arm"),
-                right_arm_move_group=self.robot_view.get_move_group("right_arm"),
-                models_dir=self.exp_config.robot_config.get_robot_dir() / "policies",
+        # The low-level WBC/PD controller (components/controller_g1ms.py) is
+        # owned here, not by whichever policy attaches via env.set_agent() --
+        # matches molmo_spaces/robots/abstract.py's Robot owning its own
+        # control stack. `low_level` lets env_g1ms.py's _load_scene carry an
+        # existing instance across scene reloads (G1Robot itself is rebuilt
+        # every reload; reusing it avoids re-loading the groot_balance/
+        # groot_walk ONNX sessions and matches how env.agent was never
+        # reconstructed on reload either) -- setup() below rebinds its
+        # qpos/qdof/actuator index arrays to this reload's model/data either
+        # way, so a fresh instance and a reused one end up equivalent.
+        if low_level is None:
+            # Deferred import -- controller_g1ms.py imports JOINT_NAMES/
+            # DEFAULT_QPOS from this module, so a module-level import here
+            # would be circular.
+            from molmo_spaces.g1_molmo_port.components.controller_g1ms import (
+                G1Controller as _LowLevelController,
             )
 
-        self._controllers = {
-            "legs_waist": legs_waist_controller,
-            "left_arm": JointPosController(self.robot_view.get_move_group("left_arm")),
-            "right_arm": JointPosController(self.robot_view.get_move_group("right_arm")),
-            "right_gripper": JointPosController(self.robot_view.get_move_group("right_gripper")),
-        }
-        assert set(self._controllers.keys()).issubset(set(self._robot_view.move_group_ids())), (
-            "All controller keys must be move group IDs"
+            low_level = _LowLevelController()
+        self._low_level = low_level
+        self._low_level.setup(model, data, prefix=self._namespace)
+
+        self._body_id = mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_BODY, self._root_body)
+        if self._body_id < 0:
+            raise RuntimeError(f"G1 root body '{self._root_body}' not found")
+        self._freejoint_id = mujoco.mj_name2id(
+            model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}floating_base_joint"
         )
 
-        # kinematics_wbc() state -- lazily built on first use (grasp-planning-
-        # specific, irrelevant to a freshly constructed robot with no grasp
-        # policy attached yet). See _ensure_wbc_ik_setup.
+        # kinematics() state -- lazily-built mink.Configuration cache (see
+        # kinematics' own docstring) plus the full scene joint-name list
+        # _solve_ik reads back solved qpos through.
+        self._stj = [
+            mujoco.mj_id2name(model, mujoco.mjtObj.mjOBJ_JOINT, i) for i in range(model.njnt)
+        ]
+        self._fj_scene = int(model.jnt_qposadr[self._freejoint_id])
+        self._ik_cfg = None
+        self._ik_mdl = None
+        self._ik_fj_qa = None
+
+        # kinematics_wbc() state -- lazily-built standalone-model mink setup,
+        # ported from agents/policy_g1ms.py's G1Controller.setup(). Built on
+        # first use rather than here since it's WBC-controller-specific
+        # (irrelevant to a G1Robot with no agent attached yet).
         self._wbc_ik_cfg = None
         self._wbc_ik_fj_dof = None
         self._wbc_ik_fj_qa = None
@@ -153,32 +370,283 @@ class G1Robot(Robot):
         self._wbc_limits = None
         self._wbc_self_collision_limit = None
 
-    @property
-    def controllers(self) -> dict[str, Controller]:
-        return self._controllers
+        self._qpos_ids = np.array(
+            [
+                model.jnt_qposadr[
+                    mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}{n}")
+                ]
+                for n in JOINT_NAMES
+            ]
+        )
+        self._dof_ids = np.array(
+            [
+                model.jnt_dofadr[
+                    mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}{n}")
+                ]
+                for n in JOINT_NAMES
+            ]
+        )
 
-    @property
-    def namespace(self):
-        return self._namespace
+        def _find_act(jname):
+            act_name = ACTUATOR_NAME_MAP.get(jname, jname)
+            aid = mujoco.mj_name2id(
+                model, mujoco.mjtObj.mjOBJ_ACTUATOR, f"{self._namespace}walk_{act_name}"
+            )
+            if aid < 0:
+                aid = mujoco.mj_name2id(
+                    model, mujoco.mjtObj.mjOBJ_ACTUATOR, f"{self._namespace}{act_name}"
+                )
+            return aid
+
+        self.act_ids = np.array([_find_act(n) for n in JOINT_NAMES])
+
+        self.right_gripper_aid = mujoco.mj_name2id(
+            model, mujoco.mjtObj.mjOBJ_ACTUATOR, f"{self._namespace}right_grip"
+        )
+        self.left_gripper_aid = mujoco.mj_name2id(
+            model, mujoco.mjtObj.mjOBJ_ACTUATOR, f"{self._namespace}left_grip"
+        )
+
+        self.n_substeps = max(1, round(0.02 / model.opt.timestep))
+
+        # Visibility uses the egocentric head camera so checks match policy POV.
+        self._cam_ids = [
+            mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_CAMERA, f"{self._namespace}{n}")
+            for n in ("head_pov",)
+            if mujoco.mj_name2id(model, mujoco.mjtObj.mjOBJ_CAMERA, f"{self._namespace}{n}") >= 0
+        ]
+        self._robot_view = G1RobotView(
+            model,
+            data,
+            self._body_id,
+            self._freejoint_id,
+            self._cam_ids,
+            self._dof_ids,
+            namespace=self._namespace,
+        )
+
+        self._apply_solver_overrides()
+        self._fix_contacts()
 
     @property
     def robot_view(self):
         return self._robot_view
 
     @property
-    def kinematics(self):
-        return self._kinematics
+    def namespace(self):
+        return self._namespace
 
     @property
-    def parallel_kinematics(self):
-        return self._parallel_kinematics
+    def controllers(self):
+        """{move_group_name: Controller}, matching molmo_spaces/robots/
+        abstract.py's Robot.controllers."""
+        controllers = getattr(self._low_level, "_controllers", None)
+        if not controllers:
+            return {}
+        return {c.move_group.name: c for c in controllers}
+
+    def set_env(self, env):
+        """Wires the low-level controller to `env` (mj model opt overrides,
+        n_substeps, floor friction -- see controller_g1ms.G1Controller.
+        set_env). Called by env_g1ms.py's _load_scene once `self` is
+        assigned to `env.robot` (set_env reads env.robot.n_substeps, so it
+        can't run from inside __init__, before that assignment exists)."""
+        self._env = env
+        self._low_level.set_env(env)
+
+    def execute_action(self, action):
+        return self._low_level.execute_action(action)
+
+    @staticmethod
+    def _floored_clip(value: float, min_mag: float, max_mag: float, deadband: float) -> float:
+        """Proportional clip, but floor the magnitude at `min_mag` once |value|
+        exceeds `deadband` -- a pure proportional-to-zero law stalls the WBC's
+        stand/walk switch (norm(cmd) < 0.05) just short of convergence."""
+        if abs(value) <= deadband:
+            return 0.0
+        return float(np.sign(value)) * float(np.clip(abs(value), min_mag, max_mag))
+
+    def waypoint_to_velocity_target(self, waypoint) -> np.ndarray:
+        """Absolute [x, y, yaw] waypoint -> the WBC's [vx, vy, yaw_rate] base
+        command, in the robot's own body frame. Same bridge (and same
+        constants) as native molmo_spaces/robots/g1.py, which is what lets a
+        molmo_spaces navigation policy drive this robot: the reference stack
+        has no such bridge because its policy emits base velocities directly.
+        """
+        xy = self._robot_view.get_xy()
+        yaw = self._robot_view.get_yaw()
+
+        dx, dy = waypoint[0] - xy[0], waypoint[1] - xy[1]
+        local_vx = np.cos(yaw) * dx + np.sin(yaw) * dy
+        local_vy = -np.sin(yaw) * dx + np.cos(yaw) * dy
+        yaw_error = normalize_ang_error(waypoint[2] - yaw)
+
+        yaw_rate = self._floored_clip(yaw_error, MIN_YAW_RATE, MAX_YAW_RATE, VELOCITY_DEADBAND)
+        if abs(yaw_error) > YAW_GATE_THRESHOLD:
+            # Turn-then-drive: while heading is substantially off, correcting it
+            # and translating at once fights itself (the gait's own turning drift
+            # keeps re-triggering a position correction that never converges).
+            vx = vy = 0.0
+        else:
+            vx = self._floored_clip(local_vx, MIN_LINEAR_VEL, MAX_LINEAR_VEL, VELOCITY_DEADBAND)
+            vy = self._floored_clip(local_vy, MIN_LINEAR_VEL, MAX_LINEAR_VEL, VELOCITY_DEADBAND)
+        return np.array([vx, vy, yaw_rate], dtype=np.float32)
+
+    def advance_control_clock(self) -> None:
+        """Advance the WBC's decimation clock by one tick.
+
+        The walking policy only re-runs its ONNX inference every 4th tick, off
+        `G1Controller._step_counter`. The reference stack increments that from
+        its *policy* (agents/policy_g1ms.py), not from execute_action, because
+        the policy is what owns the control loop -- so anything else driving
+        this robot (a navigation policy, a scripted waypoint loop) has to
+        advance it too, or the gait clock never ticks. Exposed as a real method
+        rather than having callers poke `_low_level._step_counter`.
+        """
+        self._low_level._step_counter += 1
+
+    def nav_action(self, waypoint, height=None) -> np.ndarray:
+        """A full flat-15 action that walks toward `waypoint` while holding the
+        current arm/gripper pose -- the navigation counterpart to the grasp
+        policy's own action assembly (see controller_g1ms.ACTION_DIM's layout:
+        [vx, vy, yaw_rate, height, waist(3), right_arm(7), gripper])."""
+        action = np.zeros(15, dtype=np.float32)
+        action[0:3] = self.waypoint_to_velocity_target(waypoint)
+        action[3] = float(self._low_level._height_cmd if height is None else height)
+        # waist stays 0 (level torso); arm/gripper hold their live pose so
+        # navigating never disturbs whatever the arm was doing.
+        q = self.data.qpos
+        action[7:14] = q[self._qpos_ids[22:29]]
+        action[14] = q[self._qpos_ids[29]]
+        return action
+
+    def _set_groot_defaults(self):
+        return self._low_level._set_groot_defaults()
+
+    def solve_scene_ik(
+        self, pos, rot=None, hand="right", ik_joints=None, col_limit=None, use_height=True
+    ):
+        """Full-scene mink IK. Named `solve_scene_ik` rather than `kinematics`
+        because molmo_spaces/robots/abstract.py's Robot reserves `kinematics`
+        for a *property* returning the robot's kinematics solver
+        (MlSpacesKinematics); this is a method that performs a solve. Only
+        caller is agents/policy_g1ms.py's GraspPolicy._solve_ik.
+
+        Exact relocation of agents/policy_g1ms.py's GraspPolicy._solve_ik
+        onto G1Robot -- same statements, same order, same every early-break,
+        just reading self.model/self.data (the SAME live scene MjModel/
+        MjData GraspPolicy.setup() was separately handed -- not a
+        standalone robot-only model; gold's own _solve_ik really does mink
+        IK against the whole scene) instead of GraspPolicy's own copies, and
+        taking as explicit arguments the two things that were caller
+        (GraspPolicy)-specific rather than robot-generic:
+        `ik_joints` (arm+waist for whichever hand is solving -- caller
+        computes this from its own hand/site tables) and `col_limit` (an
+        optional mink.CollisionAvoidanceLimit built by
+        GraspPolicy._build_collision_limit, which stays caller-side since
+        it's grasp-planning-specific, not robot-generic). `use_height`
+        mirrors GraspPolicy._use_height (the WBC controller sets this False
+        on its own grasp planner; the standalone grasp planner leaves it
+        True).
+
+        Deliberately NOT a property (unlike molmo_spaces' Robot.kinematics,
+        which returns a solver object) -- this is the exact callable shape
+        _solve_ik itself has, ported verbatim rather than reshaped to that
+        contract.
+
+        Returns {unprefixed_joint_name: qpos_value} for every joint in the
+        WHOLE scene model (not just this robot's) -- callers filter to the
+        ones they need, matching gold's own behavior (see _solve_ik's
+        original output loop, which never actually filtered by prefix
+        either).
+        """
+        if ik_joints is None:
+            ik_joints = set(HAND_ARM_JOINTS[hand]) | set(WAIST_JOINTS)
+        site = HAND_SITE[hand]
+
+        if self._ik_cfg is None:
+            self._ik_cfg = mink.Configuration(self.model)
+            self._ik_mdl = self._ik_cfg.model
+            self._ik_fj_qa = self._ik_mdl.jnt_qposadr[
+                mujoco.mj_name2id(
+                    self._ik_mdl, mujoco.mjtObj.mjOBJ_JOINT, self._namespace + "floating_base_joint"
+                )
+            ]
+
+        config = self._ik_cfg
+        config.update(self.data.qpos.copy())
+
+        mask = np.zeros(self._ik_mdl.nv)
+        for jn in ik_joints:
+            jid = mujoco.mj_name2id(self._ik_mdl, mujoco.mjtObj.mjOBJ_JOINT, self._namespace + jn)
+            if jid >= 0:
+                mask[self._ik_mdl.jnt_dofadr[jid]] = 1.0
+        if use_height:
+            fj_dof = self._ik_mdl.jnt_dofadr[
+                mujoco.mj_name2id(
+                    self._ik_mdl, mujoco.mjtObj.mjOBJ_JOINT, self._namespace + "floating_base_joint"
+                )
+            ]
+            mask[fj_dof + 2] = 1.0
+
+        ht = mink.FrameTask(
+            frame_name=self._namespace + site,
+            frame_type="site",
+            position_cost=100,
+            orientation_cost=1,
+            lm_damping=1,
+        )
+        post = mink.PostureTask(self._ik_mdl, cost=1e-2)
+        post.set_target_from_configuration(config)
+        r = mink.SO3.from_matrix(rot) if rot is not None else mink.SO3.identity()
+        ht.set_target(mink.SE3.from_rotation_and_translation(r, np.asarray(pos, dtype=np.float64)))
+
+        limits = [mink.ConfigurationLimit(self._ik_mdl)]
+        if col_limit is not None:
+            limits.append(col_limit)
+        prev = float("inf")
+        for step in range(300):
+            try:
+                vel = mink.solve_ik(config, [ht, post], IK_DT, "daqp", damping=1e-1, limits=limits)
+            except Exception:
+                break
+            vel *= mask
+            config.integrate_inplace(vel, IK_DT)
+            q = config.q.copy()
+            q[self._ik_fj_qa + 2] = np.clip(q[self._ik_fj_qa + 2], HEIGHT_MIN, HEIGHT_MAX)
+            config.update(q)
+            err = np.linalg.norm(ht.compute_error(config)[:3])
+            if err < 0.001:
+                break
+            if step == 100 and err > 0.1:
+                break
+            if step > 100 and err > prev - 1e-5:
+                break
+            prev = err
+
+        if use_height:
+            ik_h = np.clip(config.q[self._ik_fj_qa + 2], HEIGHT_MIN, HEIGHT_MAX)
+            self.model.qpos_spring[self._fj_scene + 2] = ik_h
+            self.model.dof_damping[self.model.jnt_dofadr[self._freejoint_id] + 2] = (
+                IK_HEIGHT_DAMPING
+            )
+
+        out = {}
+        for jn in self._stj:
+            if jn == self._namespace + "floating_base_joint":
+                continue
+            jid = mujoco.mj_name2id(self._ik_mdl, mujoco.mjtObj.mjOBJ_JOINT, jn)
+            if jid < 0:
+                continue
+            key = jn[len(self._namespace) :] if jn.startswith(self._namespace) else jn
+            out[key] = config.q[self._ik_mdl.jnt_qposadr[jid]]
+        return out
 
     def _build_wbc_self_collision_limit(self, ik_model):
-        """Self-collision limit for kinematics_wbc: arm<->torso/pelvis/waist/
-        hip + arm<->arm geom pairs. Returns None if no collidable pairs exist.
-        Moved here from FetchmanPickPlannerPolicy._build_self_collision_limit
-        -- same statements, so grasp planning stays behaviorally unaffected
-        by this move."""
+        """Exact relocation of G1Controller._build_self_collision_limit onto
+        G1Robot -- self-collision limit for kinematics_wbc: arm<->torso/
+        pelvis/waist/hip + arm<->arm geom pairs. Returns None if no
+        collidable pairs exist."""
         right_arm, left_arm, body = [], [], []
         for gid in range(ik_model.ngeom):
             if ik_model.geom_contype[gid] == 0 and ik_model.geom_conaffinity[gid] == 0:
@@ -205,28 +673,16 @@ class G1Robot(Robot):
         )
 
     def _ensure_wbc_ik_setup(self):
-        """Lazily build the standalone-model mink whole-body IK setup
-        kinematics_wbc() needs. Moved here from FetchmanPickPlannerPolicy.
-        __init__'s own IK-setup block (same statements, same order) so grasp
-        IK lives on the robot instead of being reimplemented per-policy --
-        solving mink IK against a standalone ~35-DOF model instead of the
-        live scene model (1000+ DOF in a cluttered procthor house) measured
-        ~2600x faster per iteration, which is why this doesn't just reuse
-        self.kinematics/self.mj_model directly.
-        """
+        """Lazily build the standalone-model mink setup kinematics_wbc needs,
+        exact relocation of G1Controller.setup()'s own WBC-IK construction
+        block (agents/policy_g1ms.py). Built once per G1Robot instance (this
+        robot is itself reconstructed on every scene (re)load -- see
+        env_g1ms.py's _load_scene -- matching the old per-setup()-call
+        lifetime the ported code had)."""
         if self._wbc_ik_cfg is not None:
             return
-        ik_model = mujoco.MjModel.from_xml_path(
-            str(self.exp_config.robot_config.get_robot_xml_path())
-        )
-        # Torso-tilt safety envelope, tighter than the MJCF's raw joint
-        # ranges -- must be applied before mink.Configuration(ik_model) below.
-        for jn, lo, hi in _WBC_TORSO_TILT_RANGES:
-            jid_ = mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, jn)
-            if jid_ >= 0:
-                ik_model.jnt_range[jid_] = [lo, hi]
-
-        self._wbc_ik_cfg = mink.Configuration(ik_model)
+        self._wbc_ik_cfg = mink.Configuration(mujoco.MjModel.from_xml_path(self._xml_path))
+        ik_model = self._wbc_ik_cfg.model
         ik_fj = mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, "floating_base_joint")
         self._wbc_ik_fj_dof = ik_model.jnt_dofadr[ik_fj]
         self._wbc_ik_fj_qa = ik_model.jnt_qposadr[ik_fj]
@@ -236,21 +692,25 @@ class G1Robot(Robot):
             if not jname:
                 continue
             scene_jid = mujoco.mj_name2id(
-                self.mj_model, mujoco.mjtObj.mjOBJ_JOINT, self.namespace + jname
+                self.model, mujoco.mjtObj.mjOBJ_JOINT, self._namespace + jname
             )
             if scene_jid < 0:
                 continue
             qsz = 7 if ik_model.jnt_type[jid] == mujoco.mjtJoint.mjJNT_FREE else 1
             self._wbc_scene_to_ik_qpos.append(
-                (self.mj_model.jnt_qposadr[scene_jid], ik_model.jnt_qposadr[jid], qsz)
+                (
+                    self.model.jnt_qposadr[scene_jid],
+                    ik_model.jnt_qposadr[jid],
+                    qsz,
+                )
             )
 
         self._wbc_hand_cfg = {}
         for hand in ("right", "left"):
-            arm_joints = list(_WBC_HAND_ARM_JOINTS[hand])
-            site = _WBC_HAND_SITE[hand]
+            arm_joints = list(HAND_ARM_JOINTS[hand])
+            site = HAND_SITE[hand]
             mask = np.zeros(ik_model.nv)
-            for jn in arm_joints + list(_WBC_WAIST_JOINTS):
+            for jn in arm_joints + list(WAIST_JOINTS):
                 jid = mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, jn)
                 if jid >= 0:
                     mask[ik_model.jnt_dofadr[jid]] = 1.0
@@ -279,7 +739,7 @@ class G1Robot(Robot):
             }
 
         posture_cost = np.full(ik_model.nv, 0.1)
-        for jn in _WBC_WAIST_JOINTS:
+        for jn in WAIST_JOINTS:
             jid = mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, jn)
             if jid >= 0:
                 posture_cost[ik_model.jnt_dofadr[jid]] = 0.2
@@ -287,7 +747,7 @@ class G1Robot(Robot):
         self._wbc_waist_qa = np.array(
             [
                 ik_model.jnt_qposadr[mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, jn)]
-                for jn in _WBC_WAIST_JOINTS
+                for jn in WAIST_JOINTS
             ],
             dtype=np.int32,
         )
@@ -299,32 +759,40 @@ class G1Robot(Robot):
             orientation_cost=0,
             lm_damping=1,
         )
+        for jn, lo, hi in [
+            ("waist_yaw_joint", -0.5, 0.5),
+            ("waist_roll_joint", -0.4, 0.4),
+            ("waist_pitch_joint", -0.1, 0.4),
+        ]:
+            jid = mujoco.mj_name2id(ik_model, mujoco.mjtObj.mjOBJ_JOINT, jn)
+            if jid >= 0:
+                ik_model.jnt_range[jid] = [lo, hi]
         self._wbc_limits = [mink.ConfigurationLimit(ik_model)]
         self._wbc_self_collision_limit = self._build_wbc_self_collision_limit(ik_model)
 
     def kinematics_wbc(
         self, target_pos, target_rot=None, hand="right", avoid_self_collision=False, precision=False
     ):
-        """Whole-body grasp IK: solves for arm+waist+height that places `hand`'s
-        gripper site at `target_pos`/`target_rot`, against a standalone
-        robot-only model synced from the live scene qpos through a
-        precomputed joint correspondence table. Moved here from
-        FetchmanPickPlannerPolicy._solve_ik_wbc (same statements, same order,
-        same every early-break) so grasp-reach IK lives on the robot instead
-        of being duplicated per-policy. `precision` replaces the caller's own
-        `self._grasp_phase in (...)` check (that's grasp-policy state, not
-        robot-generic) -- True selects the tighter max_iters/conv_thresh used
-        during PHASE_DESCEND/CLOSE/POST_CLOSE/LIFT.
+        """Exact relocation of agents/policy_g1ms.py's G1Controller.
+        _solve_ik_wbc onto G1Robot -- same statements, same order, same
+        every early-break. Unlike kinematics() above, this solves against a
+        standalone robot-only model (mirrors G1Controller.setup()'s own
+        `mujoco.MjModel.from_xml_path(self._xml_path)`), synced from the live
+        scene qpos through a precomputed joint correspondence table rather
+        than operating on the scene model directly. `precision` replaces
+        the caller's own `self._grasp_phase in (...)` check (the caller
+        computes that condition, since it's WBC-controller-state-specific,
+        not robot-generic) -- True selects the tighter max_iters/conv_thresh
+        gold used during PHASE_DESCEND/CLOSE/POST_CLOSE/LIFT.
 
-        Returns (arm, waist, ik_h, err): arm joint qpos (7,), waist joint
-        qpos (3,), clamped pelvis height, final position error norm.
+        Returns (arm, waist, ik_h, err) exactly as _solve_ik_wbc did.
         """
         self._ensure_wbc_ik_setup()
         hcfg = self._wbc_hand_cfg[hand]
 
         ik_q = np.zeros(self._wbc_ik_cfg.model.nq, dtype=np.float64)
         for scene_qa, ik_qa, qsz in self._wbc_scene_to_ik_qpos:
-            ik_q[ik_qa : ik_qa + qsz] = self.mj_data.qpos[scene_qa : scene_qa + qsz]
+            ik_q[ik_qa : ik_qa + qsz] = self.data.qpos[scene_qa : scene_qa + qsz]
         self._wbc_ik_cfg.update(ik_q)
 
         q_post = self._wbc_ik_cfg.q.copy()
@@ -334,7 +802,7 @@ class G1Robot(Robot):
         # so the IK actively wants to stand back up when the wrist target allows it.
         pelvis_T = self._wbc_ik_cfg.get_transform_frame_to_world("pelvis", "body")
         pelvis_pos = pelvis_T.translation().copy()
-        pelvis_pos[2] = _WBC_HEIGHT_MAX
+        pelvis_pos[2] = HEIGHT_MAX
         self._wbc_pelvis_task.set_target(
             mink.SE3.from_rotation_and_translation(pelvis_T.rotation(), pelvis_pos)
         )
@@ -366,7 +834,7 @@ class G1Robot(Robot):
             self._wbc_ik_cfg.integrate_inplace(vel, 1e-2)
             q_tmp = self._wbc_ik_cfg.q.copy()
             q_tmp[self._wbc_ik_fj_qa + 2] = np.clip(
-                q_tmp[self._wbc_ik_fj_qa + 2], _WBC_HEIGHT_MIN, _WBC_HEIGHT_MAX
+                q_tmp[self._wbc_ik_fj_qa + 2], HEIGHT_MIN, HEIGHT_MAX
             )
             self._wbc_ik_cfg.update(q_tmp)
             err = float(np.linalg.norm(hcfg["task"].compute_error(self._wbc_ik_cfg)[:3]))
@@ -378,261 +846,172 @@ class G1Robot(Robot):
         q = self._wbc_ik_cfg.q
         arm = q[hcfg["arm_qa"]].astype(np.float32)
         waist = q[self._wbc_waist_qa].astype(np.float32)
-        ik_h = float(np.clip(q[self._wbc_ik_fj_qa + 2], _WBC_HEIGHT_MIN, _WBC_HEIGHT_MAX))
+        ik_h = float(np.clip(q[self._wbc_ik_fj_qa + 2], HEIGHT_MIN, HEIGHT_MAX))
         return arm, waist, ik_h, err
 
-    def get_arm_move_group_ids(self) -> list[str]:
-        """Both arms get independent TCP-bounded action noise."""
-        return ["left_arm", "right_arm"]
+    @property
+    def kinematics(self):
+        """Robot ABC's kinematics solver. Built lazily: the reference stack
+        constructs this robot with no exp_config at all (it drives the arm
+        through solve_scene_ik/kinematics_wbc, never through
+        MlSpacesKinematics), so requiring one up front would break that path.
+        Native callers that do pass an exp_config get the same solver every
+        other molmo_spaces Robot exposes."""
+        if self._kinematics is None:
+            self._kinematics = MlSpacesKinematics(self._require_exp_config("kinematics"))
+        return self._kinematics
 
-    def update_control(self, action_command_dict: dict[str, Any]) -> None:
-        """Bridge nav policies' base actions to whichever base control mode is active.
+    @property
+    def parallel_kinematics(self):
+        """Robot ABC's parallel kinematics. DummyParallelKinematics matches
+        what native G1Robot uses -- G1 has no batched-IK backend."""
+        if self._parallel_kinematics is None:
+            self._parallel_kinematics = DummyParallelKinematics(
+                self._require_exp_config("parallel_kinematics"), self.kinematics
+            )
+        return self._parallel_kinematics
 
-        Nav policies built for holonomic/mocap-driven bases (see
-        AStarPlannerPolicy._build_navigation_action) command an absolute
-        world-frame [x, y, theta] waypoint under the "base" key -- but G1 has no
-        "base" move group controller in either mode (walking: the pelvis has no
-        actuators at all; holo: the mocap target is written directly in
-        compute_control(), not via the Controller/set_target() flow). Translate
-        the waypoint into whichever interface is active before delegating to the
-        normal per-move-group update_control() flow.
+    def _require_exp_config(self, what: str):
+        if self.exp_config is None:
+            raise RuntimeError(
+                f"G1Robot.{what} needs an exp_config, but this robot was constructed "
+                "without one (the g1_molmo_port stack builds it straight from "
+                "model/data). Construct it with exp_config= to use this."
+            )
+        return self.exp_config.robot_config
 
-        FetchManBasePlannerPolicy instead computes a [vx, vy, yaw_rate] velocity
-        command itself every step (see its _update_nav_command) and sends it
-        directly under "base_velocity", bypassing _waypoint_to_velocity_target
-        entirely -- there's no waypoint pose to re-derive it from. WBC mode only;
-        holo-base mode has no notion of a velocity command (its mocap target is
-        an absolute pose), so "base_velocity" is ignored there.
-        """
-        action_command_dict = dict(action_command_dict)
-        waypoint = action_command_dict.pop("base", None)
-        base_velocity = action_command_dict.pop("base_velocity", None)
-        # G1BaseGroup.noop_ctrl is an empty array (the pelvis has no actuators), so
-        # e.g. AStarPlannerPolicy._build_done_action's {"base": get_noop_ctrl_dict(...)}
-        # sends an empty array here rather than omitting the key or sending None.
-        has_waypoint = waypoint is not None and len(waypoint) == 3
-        if self._use_holo_base:
-            self._pending_base_ctrl = (
-                self._waypoint_to_pose_ctrl(waypoint) if has_waypoint else None
-            )
-        elif has_waypoint:
-            action_command_dict["legs_waist"] = self._waypoint_to_velocity_target(waypoint)
+    def _apply_solver_overrides(self):
+        m = self.model
+        m.opt.integrator = mujoco.mjtIntegrator.mjINT_IMPLICITFAST
+        m.opt.cone = mujoco.mjtCone.mjCONE_PYRAMIDAL
+        m.opt.noslip_iterations = 5
+        m.opt.gravity[:] = [0, 0, -9.81]
+        m.opt.impratio = 1.0
+        m.opt.jacobian = 2  # auto
+        m.opt.enableflags = 0
+        for gid in range(m.ngeom):
+            gname = mujoco.mj_id2name(m, mujoco.mjtObj.mjOBJ_GEOM, gid) or ""
+            if _is_floor_geom(gname):
+                m.geom_friction[gid] = [1.0, 0.005, 0.0001]
 
-        if base_velocity is not None and not self._use_holo_base:
-            # Apply the same floored-clip treatment _waypoint_to_velocity_target
-            # already does for the "base" waypoint path -- without it, a nav
-            # policy's own smoothstep brake (ramping speed continuously down to
-            # 0 on final approach, e.g. FetchManBasePlannerPolicy/
-            # FetchmanPickPlannerPolicy's _update_nav_command, ported directly
-            # from g1_molmo) spends real time commanding small-but-nonzero
-            # speeds in [_VELOCITY_DEADBAND, _MIN_LINEAR_VEL) -- exactly the
-            # regime _floored_clip's docstring describes G1WalkController's
-            # stand/walk switch getting stuck in, since a raw pass-through
-            # command like this bypassed the fix entirely. Confirmed
-            # empirically: FetchmanPickPlannerPolicy's walk phase stalled
-            # ~0.28m short of its goal indefinitely, sending a steady ~0.09
-            # m/s (between the 0.08 deadband and the 0.15 floor) the whole
-            # time with essentially zero actual displacement.
-            vx = self._floored_clip(
-                base_velocity[0], _MIN_LINEAR_VEL, _MAX_LINEAR_VEL, _VELOCITY_DEADBAND
-            )
-            vy = self._floored_clip(
-                base_velocity[1], _MIN_LINEAR_VEL, _MAX_LINEAR_VEL, _VELOCITY_DEADBAND
-            )
-            yaw_rate = self._floored_clip(
-                base_velocity[2], _MIN_YAW_RATE, _MAX_YAW_RATE, _VELOCITY_DEADBAND
-            )
-            action_command_dict["legs_waist"] = np.array(
-                [vx, vy, yaw_rate, _DEFAULT_HEIGHT_CMD, 0.0, 0.0, 0.0], dtype=np.float32
-            )
+    def _fix_contacts(self):
+        m = self.model
+        for gid in range(m.ngeom):
+            bid = m.geom_bodyid[gid]
+            bname = mujoco.mj_id2name(m, mujoco.mjtObj.mjOBJ_BODY, bid) or ""
+            if bname.startswith(self._namespace) and "ankle_roll" in bname:
+                if m.geom_type[gid] == mujoco.mjtGeom.mjGEOM_SPHERE:
+                    m.geom_conaffinity[gid] = 15
 
-        legs_waist_action = action_command_dict.get("legs_waist")
-        if (
-            not self._use_holo_base
-            and legs_waist_action is not None
-            and len(legs_waist_action) != NUM_TARGET_DIMS
+    def set_defaults(self):
+        self.data.qpos[self._qpos_ids] = DEFAULT_QPOS
+        valid = self.act_ids >= 0
+        self.data.ctrl[self.act_ids[valid]] = DEFAULT_QPOS[valid]
+        for i, name in enumerate(JOINT_NAMES):
+            for prefix in ("walk_", "grasp_"):
+                aid = mujoco.mj_name2id(
+                    self.model, mujoco.mjtObj.mjOBJ_ACTUATOR, f"{self._namespace}{prefix}{name}"
+                )
+                if aid >= 0:
+                    self.data.ctrl[aid] = DEFAULT_QPOS[i]
+        if self.right_gripper_aid >= 0:
+            self.data.ctrl[self.right_gripper_aid] = GRIPPER_OPEN
+        for jname, qval in (
+            ("right_Joint1_1", GRIPPER_OPEN),
+            ("right_Joint2_1", GRIPPER_OPEN),
         ):
-            # Generic policies unaware of G1's WBC (e.g. base_object_manipulation_
-            # planner_policy's get_noop_ctrl_dict(), used to hold non-manipulated
-            # move groups still during e.g. door/cabinet opening) send legs_waist's
-            # own MoveGroup-shaped noop ctrl -- a 15-dim joint-position array,
-            # matching its actuator count -- not realizing G1WalkController's
-            # target is a 7-dim [vx, vy, yaw_rate, height, waist] velocity command.
-            # Drop it so update_control()'s normal "no command -> set_to_stationary()"
-            # fallback applies instead, which is what "hold legs_waist still"
-            # actually means for a WBC-driven biped (see G1WalkController.
-            # set_to_stationary's docstring).
-            action_command_dict.pop("legs_waist")
-
-        super().update_control(action_command_dict)
-
-    def compute_control(self) -> None:
-        super().compute_control()
-        if self._use_holo_base:
-            base = self._robot_view.get_move_group("base")
-            base.ctrl = (
-                self._pending_base_ctrl if self._pending_base_ctrl is not None else (base.noop_ctrl)
+            jid = mujoco.mj_name2id(
+                self.model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}{jname}"
             )
+            if jid >= 0:
+                self.data.qpos[self.model.jnt_qposadr[jid]] = qval
 
-    @staticmethod
-    def _floored_clip(value: float, min_mag: float, max_mag: float, deadband: float) -> float:
-        """Proportional clip, but floor the magnitude at `min_mag` once |value| exceeds
-        `deadband` -- see _VELOCITY_DEADBAND for why a pure proportional-to-zero law
-        stalls G1WalkController's stand/walk switch short of convergence."""
-        if abs(value) <= deadband:
-            return 0.0
-        return float(np.sign(value)) * float(np.clip(abs(value), min_mag, max_mag))
+    # waist(3) + right arm(7) + right Joint1_1(1).
+    _UPPER_RAND_IDX = np.array([12, 13, 14, 22, 23, 24, 25, 26, 27, 28, 29], dtype=np.int64)
+    _UPPER_GRIPPER_LOCAL = 10
 
-    def _waypoint_to_velocity_target(self, waypoint) -> np.ndarray:
-        pose = self._robot_view.base.pose
-        x, y = pose[0, 3], pose[1, 3]
-        yaw = R.from_matrix(pose[:3, :3]).as_euler("xyz")[2]
+    def sample_upper_pose(self, np_random, radius=0.15):
+        idx = self._UPPER_RAND_IDX
+        # Waist gets 0.4x the arm's noise range so the trunk stays closer to neutral.
+        per_dim = np.full(len(idx), radius, dtype=np.float64)
+        per_dim[:3] = radius * 0.4
+        sampled = DEFAULT_QPOS[idx] + np_random.uniform(-per_dim, per_dim)
+        sampled[self._UPPER_GRIPPER_LOCAL] = float(np_random.uniform(GRIPPER_OPEN, GRIPPER_CLOSED))
+        for i, jidx in enumerate(idx):
+            jid = mujoco.mj_name2id(
+                self.model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}{JOINT_NAMES[jidx]}"
+            )
+            if jid >= 0:
+                lo, hi = self.model.jnt_range[jid]
+                sampled[i] = float(np.clip(sampled[i], lo, hi))
+        return sampled
 
-        dx, dy = waypoint[0] - x, waypoint[1] - y
-        local_vx = np.cos(yaw) * dx + np.sin(yaw) * dy
-        local_vy = -np.sin(yaw) * dx + np.cos(yaw) * dy
-        yaw_error = normalize_ang_error(waypoint[2] - yaw)
-
-        yaw_rate = self._floored_clip(yaw_error, _MIN_YAW_RATE, _MAX_YAW_RATE, _VELOCITY_DEADBAND)
-        if abs(yaw_error) > _YAW_GATE_THRESHOLD:
-            # Turn-then-drive: while heading is substantially off, correcting it and
-            # translating simultaneously fights itself (the walking gait's own
-            # turning drift keeps re-triggering a position correction, which never
-            # converges -- confirmed empirically via rotate() plateaux). Prioritize
-            # closing the heading error first; translate once roughly facing the
-            # waypoint, same as a differential-drive "rotate then drive" strategy.
-            vx = vy = 0.0
-        else:
-            vx = self._floored_clip(local_vx, _MIN_LINEAR_VEL, _MAX_LINEAR_VEL, _VELOCITY_DEADBAND)
-            vy = self._floored_clip(local_vy, _MIN_LINEAR_VEL, _MAX_LINEAR_VEL, _VELOCITY_DEADBAND)
-
-        return np.array([vx, vy, yaw_rate, _DEFAULT_HEIGHT_CMD, 0.0, 0.0, 0.0], dtype=np.float32)
-
-    def _waypoint_to_pose_ctrl(self, waypoint) -> np.ndarray:
-        """Holo-base mode: convert an absolute [x, y, theta] waypoint directly into
-        the mocap target's [x, y, z, qw, qx, qy, qz] ctrl (see G1HoloBaseGroup),
-        keeping the current height and a level (roll=pitch=0) orientation."""
-        current_z = self._robot_view.base.pose[2, 3]
-        quat = R.from_euler("z", waypoint[2]).as_quat(scalar_first=True)
-        return np.array([waypoint[0], waypoint[1], current_z, *quat], dtype=np.float32)
-
-    def reset(self) -> None:
-        """Reset the robot to its initial standing state."""
-        init_qpos_dict = self.exp_config.robot_config.init_qpos
-        self.set_joint_pos(init_qpos_dict)
-        self._pending_base_ctrl = None
-        for controller in self._controllers.values():
-            controller.reset()
-
-    @staticmethod
-    def robot_model_root_name() -> str:
-        return "pelvis"
-
-    @classmethod
-    def add_robot_to_scene(
-        cls,
-        robot_config: "BaseRobotConfig",
-        spec: MjSpec,
-        prefix: str,
-        pos: list[float],
-        quat: list[float],
-        randomize_textures: bool = False,
-        strip_meshes: bool = False,
-    ) -> None:
-        pos = pos + [0.0] if len(pos) == 2 else pos
-        super().add_robot_to_scene(
-            robot_config=robot_config,
-            spec=spec,
-            prefix=prefix,
-            pos=pos,
-            quat=quat,
-            randomize_textures=randomize_textures,
-            strip_meshes=strip_meshes,
+    def apply_upper_pose(self, values):
+        idx = self._UPPER_RAND_IDX
+        self.data.qpos[self._qpos_ids[idx]] = values
+        aids = self.act_ids[idx]
+        valid = aids >= 0
+        if valid.any():
+            self.data.ctrl[aids[valid]] = values[valid]
+        # Mirror Joint2_1 to Joint1_1 — the equality only enforces during sim, not on init.
+        j2 = mujoco.mj_name2id(
+            self.model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}right_Joint2_1"
         )
+        if j2 >= 0:
+            self.data.qpos[self.model.jnt_qposadr[j2]] = values[self._UPPER_GRIPPER_LOCAL]
 
-        if getattr(robot_config, "use_holo_base", False):
-            # Mirrors FloatingRUMRobot.add_robot_to_scene: weld a mocap target body
-            # to the pelvis so G1HoloBaseGroup can drive the base directly via ctrl,
-            # since G1's pelvis is a real free joint (no virtual holonomic joints to
-            # actuate the way RBY1's base does).
-            target_body_name = f"{prefix}{HOLO_BASE_TARGET_BODY_NAME}"
-            spec.worldbody.add_body(name=target_body_name, pos=pos, quat=quat, mocap=True)
-            eq = spec.add_equality()
-            eq.name1 = target_body_name
-            eq.name2 = f"{prefix}{cls.robot_model_root_name()}"
-            eq.solref = np.array([0.02, 1])
-            eq.solimp = np.array([0.9, 0.95, 0.0, 1, 2])
-            eq.objtype = mjtObj.mjOBJ_BODY
-            eq.type = mjtEq.mjEQ_WELD
+    def apply_arm_pose(self, joints_dict):
+        """Apply a {joint_name: qpos_value} dict produced by IK to the scene qpos.
+        Only writes joints in JOINT_NAMES (skips legs/freejoint). Also mirrors
+        right_Joint2_1 to right_Joint1_1."""
+        if not joints_dict:
+            return
+        for i, jname in enumerate(JOINT_NAMES):
+            if jname in joints_dict:
+                self.data.qpos[self._qpos_ids[i]] = float(joints_dict[jname])
+                aid = int(self.act_ids[i])
+                if aid >= 0:
+                    self.data.ctrl[aid] = float(joints_dict[jname])
+        j2 = mujoco.mj_name2id(
+            self.model, mujoco.mjtObj.mjOBJ_JOINT, f"{self._namespace}right_Joint2_1"
+        )
+        if j2 >= 0 and "right_Joint1_1" in joints_dict:
+            self.data.qpos[self.model.jnt_qposadr[j2]] = float(joints_dict["right_Joint1_1"])
 
-    @classmethod
-    def apply_control_overrides(cls, spec: MjSpec, robot_config: "BaseRobotConfig"):
-        super().apply_control_overrides(spec, robot_config)
-        # Match the solver settings the source G1 stack validated for stable
-        # bipedal contact: implicit integration, pyramidal friction cones,
-        # extra no-slip iterations, and unit impedance ratio. The robot's own
-        # MJCF <option> block can't take effect on its own -- MjSpec.attach_body
-        # only merges body subtrees, not <option> blocks -- so whatever the
-        # target scene's <option> says would otherwise silently win.
-        spec.option.integrator = mujoco.mjtIntegrator.mjINT_IMPLICITFAST
-        spec.option.cone = mujoco.mjtCone.mjCONE_PYRAMIDAL
-        spec.option.noslip_iterations = 5
-        spec.option.impratio = 1.0
+    def set_pose(self, xy, yaw, z=STANDING_HEIGHT):
+        self._robot_view.set_pose(xy, yaw, z=z)
 
-        # The MJCF's "walk_*" actuators for legs_waist are authored as tuned
-        # position-PD actuators (biastype="affine"), matching Phase 2's plain
-        # JointPosController usage. G1WalkController computes its own PD torque
-        # in Python (see molmo_spaces.controllers.g1_walk), so these need to
-        # become raw torque passthroughs instead: fixed unit gain, no bias.
-        # Holo-base mode doesn't use G1WalkController (legs_waist just holds a
-        # static pose via plain JointPosController), so it keeps the MJCF's
-        # original tuned PD gains untouched.
-        namespace = robot_config.robot_namespace
-        if not getattr(robot_config, "use_holo_base", False):
-            for joint_name in LEGS_WAIST_JOINT_SUFFIXES:
-                actuator = spec.actuator(f"{namespace}walk_{joint_name}")
-                assert actuator is not None, f"Missing walk_{joint_name} actuator"
-                actuator.gaintype = mujoco.mjtGain.mjGAIN_FIXED
-                actuator.biastype = mujoco.mjtBias.mjBIAS_NONE
-                actuator.gainprm[0] = 1.0
-                actuator.biasprm[:] = 0.0
+    def zero_velocities(self):
+        self._robot_view.zero_velocities()
 
-            # The arms' own "walk_{side}_*" actuators are authored in the MJCF
-            # with much weaker gains than G1WalkController's IK targets assume
-            # (e.g. the wrist actuators cap out around +-5 N*m) -- reconfigure
-            # them to the same kp/kd the reference g1_molmo stack uses
-            # (components/controller_g1ms.py's G1Controller.setup()), or the
-            # arm/wrist can't physically reach IK-commanded orientations:
-            # position converges "close enough" while rotation error stalls
-            # at a fixed residual forever. Left arm is zero-gained (hangs
-            # passively under gravity, matching the reference) even though it
-            # still sits behind a JointPosController -- gain=0 makes whatever
-            # that controller writes moot.
-            for joint_name in ARM_JOINT_SUFFIXES:
-                left_actuator = spec.actuator(f"{namespace}walk_left_{joint_name}")
-                assert left_actuator is not None, f"Missing walk_left_{joint_name} actuator"
-                left_actuator.gainprm[0] = 0.0
-                left_actuator.biasprm[:] = 0.0
-                left_actuator.forcerange[:] = [-400, 400]
+    def get_xy(self):
+        return self._robot_view.get_xy()
 
-                right_actuator = spec.actuator(f"{namespace}walk_right_{joint_name}")
-                assert right_actuator is not None, f"Missing walk_right_{joint_name} actuator"
-                right_actuator.gainprm[0] = 2000.0
-                right_actuator.biasprm[0] = 0.0
-                right_actuator.biasprm[1] = -2000.0
-                right_actuator.biasprm[2] = -60.0
-                right_actuator.forcerange[:] = [-400, 400]
+    def get_yaw(self):
+        return self._robot_view.get_yaw()
 
-        # The MJCF puts the head/mount/logo visual geoms on group 5 (every other
-        # visual geom uses the "visual" class default of group 2). mujoco.MjvOption's
-        # default geomgroup is [1, 1, 1, 0, 0, 0] and this codebase's renderers never
-        # enable group 5 (only sitegroup gets touched -- see MjOpenGLRenderer), so the
-        # head/logo are silently invisible in every render. Move them to group 2 so
-        # they render like the rest of the robot.
-        head_mesh_suffixes = ("head_link", "head_mount", "logo_link")
-        for body_name in ("torso_link", "head_camera_mount"):
-            body = spec.body(f"{namespace}{body_name}")
-            assert body is not None, f"Missing {body_name} body"
-            for geom in body.geoms:
-                if geom.group == 5 and geom.meshname.endswith(head_mesh_suffixes):
-                    geom.group = 2
+    def pelvis_height(self):
+        return self._robot_view.pelvis_height()
+
+    def place(self, xy, yaw):
+        self.set_pose(np.asarray(xy, dtype=np.float64), float(yaw))
+        self.set_defaults()
+        self.zero_velocities()
+        mujoco.mj_forward(self.model, self.data)
+
+    def has_bad_contacts(self):
+        return self._robot_view.has_bad_contacts()
+
+    def check_object_visibility(self, body_id, threshold=0.00002):
+        return self._robot_view.check_object_visibility(body_id, threshold=threshold)
+
+    def close(self):
+        """Free the lazily-created visibility renderer (and its GL/EGL context).
+        Must be called before dropping the robot on scene reload — otherwise the
+        renderer's framebuffer leaks on the render GPU because __del__-based EGL
+        teardown is unreliable, so VRAM creeps to OOM across reloads."""
+        self._robot_view.close()
+
+    def state_is_finite(self):
+        d = self.data
+        return np.isfinite(d.qpos).all() and np.isfinite(d.qvel).all() and np.isfinite(d.ctrl).all()
